Never use built-in attribute names on OpenGL ES
They are not allowed in the specification:

From GLSL_ES_Specification_1.0.17.pdf:

"7.3 Vertex Shader Built-In Attributes

There are no built-in attribute names in OpenGL ES."

This will cause vertex shaders that use the same semantic
more than once to create two attribute variables with the
same name which will fail to compile. Subsequent commits
in this series address that.