Suppress attributes with duplicate semantics
.. otherwise you can end up with multiple attribute
variables with the same name. I determine the usage
of the semantic with the greatest number of elements
and suppress all the others.

A common base class for GlslSymbol and StructMember
(GlslSymbolOrStructMemberBase) was added to enable
this.

New test-case (in-struct-dupe-sem-ret-struct-in.txt)
was added for this feature.

Updated -outES.txt/-outES3.txt test-cases with these
now non-duplicated attributes.