fixed from PVS-Studio
V668 There is no sense in testing the 'memory' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. poolalloc.cpp 222
V668 There is no sense in testing the 'memory' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. poolalloc.cpp 246
V668 There is no sense in testing the 'lpThreadData' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. parsehelper.cpp 1918